Clean up temporary files after successfully writing a ZIP archive.
git-svn-id: https://svn.code.sf.net/p/vassalengine/svn/VASSAL-src/trunk@9017 67b53d14-2c14-4ace-a08f-0dab2b34000c

+++ src/VASSAL/tools/io/ZipArchive.java
@@ -362,7 +362,7 @@ public class ZipArchive implements FileArchive {
 
       // delete all temporary files
       for (Entry e : entries.values()) {
-        if (e.file != null) {
+        if (e != null && e.file != null) {
           e.file.delete();
         }
       }
@@ -506,6 +506,13 @@ public class ZipArchive implements FileArchive {
 
         err += " Data written to " + tmpFile.getAbsolutePath() + " instead.";
         throw (IOException) new IOException(err).initCause(e);
+      }
+    }
+
+    // Delete all temporary files
+    for (Entry e : entries.values()) {
+      if (e != null && e.file != null) {
+        e.file.delete();
       }
     }
